feat: rendering the tiles textures using the current world model.
feat: uses the new viewport component to visualize the current world.

+++ MMORPGEngine/Data/Ground/GroundCatalog.h
@@ -3,7 +3,7 @@
 
 #include <unordered_map>
 
-#include <MMORPGEngine/World/Tile/GroundModel.h>
+#include <MMORPGEngine/Data/Ground/GroundModel.h>
 
 namespace Engine {
 
@@ -11,12 +11,12 @@ class GroundCatalog {
 public:
     GroundCatalog();
 
-    const GroundModel* ground( uint16_t type ) const;
-    const std::unordered_map<uint16_t, GroundModel>& grounds() const;
+    const GroundModel* ground( uint32_t type ) const;
+    const std::unordered_map<uint32_t, GroundModel>& grounds() const;
     void addGround( const GroundModel& ground );
 
 private:
-    std::unordered_map<uint16_t, GroundModel> _grounds;
+    std::unordered_map<uint32_t, GroundModel> _grounds;
 };
 
 } // namespace Engine
